+++ modules/16bit/code/sampler/includes/audio/env/AttackRelease.h
@@ -72,7 +72,7 @@ class AttackReleaseEnvelope : public Envelope {
             }
         }
 
-        int16_t process(int16_t sample) override {
+        float process(float sample) override {
             // Check for zero-crossing from positive to negative
             if (triggerAtZero && previousSample >= 0 && sample < 0) {
                 triggerAtZero = false;
